+++ src/PKToken.m
@@ -244,7 +244,7 @@ static PKTokenEOF *EOFToken = nil;
         typeString = @"Hashtag";
 #endif
     }
-    return [NSString stringWithFormat:@"<%@ %C%@%C>", typeString, 0x00AB, self.value, 0x00BB];
+    return [NSString stringWithFormat:@"<%@ %C%@%C>", typeString, ( unsigned short )0x00AB, self.value, ( unsigned short )0x00BB];
 }
